more decoupling bewtween renderer and crafter
