+++ app/cli/cmd/workflow_workflow_run_describe.go
@@ -25,7 +25,7 @@ import (
 	"time"
 
 	"github.com/chainloop-dev/chainloop/app/cli/internal/action"
-	"github.com/chainloop-dev/chainloop/pkg/attestation/renderer"
+	"github.com/chainloop-dev/chainloop/pkg/attestation/renderer/chainloop"
 	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/jedib0t/go-pretty/v6/text"
 	"github.com/muesli/reflow/wrap"
@@ -220,7 +220,7 @@ func predicateV1Table(att *action.WorkflowRunAttestationItem) {
 		mt.Render()
 	}
 
-	evs := att.PolicyEvaluations[renderer.AttPolicyEvaluation]
+	evs := att.PolicyEvaluations[chainloop.AttPolicyEvaluation]
 	if len(evs) > 0 {
 		mt := newTableWriter()
 		mt.SetTitle("Attestation policies")
