+++ pkg/attestation/crafter/crafter_test.go
@@ -26,6 +26,7 @@ import (
 	schemaapi "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
 	"github.com/chainloop-dev/chainloop/internal/casclient"
 	mUploader "github.com/chainloop-dev/chainloop/internal/casclient/mocks"
+	"github.com/chainloop-dev/chainloop/pkg/attestation"
 	"github.com/chainloop-dev/chainloop/pkg/attestation/crafter"
 	v1 "github.com/chainloop-dev/chainloop/pkg/attestation/crafter/api/attestation/v1"
 	"github.com/chainloop-dev/chainloop/pkg/attestation/crafter/materials"
@@ -542,7 +543,7 @@ func loadSchema(path string) (*schemaapi.CraftingSchema, error) {
 		return nil, err
 	}
 
-	jsonSchemaRaw, err := materials.LoadJSONBytes(content, filepath.Ext(path))
+	jsonSchemaRaw, err := attestation.LoadJSONBytes(content, filepath.Ext(path))
 	if err != nil {
 		return nil, err
 	}
