+++ pkg/attestation/renderer/renderer.go
@@ -31,7 +31,6 @@ import (
 	v1 "github.com/chainloop-dev/chainloop/pkg/attestation/crafter/api/attestation/v1"
 	"github.com/chainloop-dev/chainloop/pkg/attestation/renderer/chainloop"
 	chainloopsigner "github.com/chainloop-dev/chainloop/pkg/attestation/signer/chainloop"
-	"github.com/chainloop-dev/chainloop/pkg/policies"
 	intoto "github.com/in-toto/attestation/go/v1"
 	"github.com/rs/zerolog"
 	"github.com/secure-systems-lab/go-securesystemslib/dsse"
@@ -41,7 +40,6 @@ import (
 	sigstoresigner "github.com/sigstore/sigstore/pkg/signature"
 	sigdsee "github.com/sigstore/sigstore/pkg/signature/dsse"
 	"google.golang.org/protobuf/encoding/protojson"
-	"google.golang.org/protobuf/types/known/structpb"
 )
 
 type AttestationRenderer struct {
@@ -56,12 +54,10 @@ type AttestationRenderer struct {
 }
 
 type r interface {
-	Statement() (*intoto.Statement, error)
+	Statement(ctx context.Context) (*intoto.Statement, error)
 }
 
 type Opt func(*AttestationRenderer)
-
-const AttPolicyEvaluation = "CHAINLOOP.ATTESTATION"
 
 func WithLogger(logger zerolog.Logger) Opt {
 	return func(ar *AttestationRenderer) {
@@ -86,13 +82,14 @@ func NewAttestationRenderer(state *crafter.VersionedCraftingState, attClient pb.
 		schema:     state.GetInputSchema(),
 		dsseSigner: sigdsee.WrapSigner(signer, "application/vnd.in-toto+json"),
 		signer:     signer,
-		renderer:   chainloop.NewChainloopRendererV02(state.GetAttestation(), builderVersion, builderDigest),
 		attClient:  attClient,
 	}
 
 	for _, opt := range opts {
 		opt(r)
 	}
+
+	r.renderer = chainloop.NewChainloopRendererV02(state.GetAttestation(), state.GetInputSchema(), builderVersion, builderDigest, attClient, &r.logger)
 
 	return r, nil
 }
@@ -102,27 +99,13 @@ func NewAttestationRenderer(state *crafter.VersionedCraftingState, attClient pb.
 func (ab *AttestationRenderer) Render(ctx context.Context) (*dsse.Envelope, error) {
 	ab.logger.Debug().Msg("generating in-toto statement")
 
-	statement, err := ab.renderer.Statement()
+	statement, err := ab.renderer.Statement(ctx)
 	if err != nil {
 		return nil, err
 	}
 
 	if err := statement.Validate(); err != nil {
 		return nil, fmt.Errorf("validating intoto statement: %w", err)
-	}
-
-	// validate attestation-level policies
-	pv := policies.NewPolicyVerifier(ab.schema, ab.attClient, &ab.logger)
-	policyResults, err := pv.VerifyStatement(ctx, statement)
-	if err != nil {
-		return nil, fmt.Errorf("applying policies to statement: %w", err)
-	}
-	// log policy violations
-	policies.LogPolicyViolations(policyResults, &ab.logger)
-
-	// insert attestation level policy results into statement
-	if err = addPolicyResults(statement, policyResults); err != nil {
-		return nil, fmt.Errorf("adding policy results to statement: %w", err)
 	}
 
 	rawStatement, err := protojson.Marshal(statement)
@@ -158,52 +141,6 @@ func (ab *AttestationRenderer) Render(ctx context.Context) (*dsse.Envelope, erro
 	}
 
 	return &dsseEnvelope, nil
-}
-
-// addPolicyResults adds policy evaluation results to the statement. It does it by deserializing the predicate from a structpb.Struct,
-// filling PolicyEvaluations, and serializing it again to a structpb.Struct object, using JSON as an intermediate representation.
-// Note that this is needed because intoto predicates are generic structpb.Struct
-func addPolicyResults(statement *intoto.Statement, policyResults []*v1.PolicyEvaluation) error {
-	predicate := statement.Predicate
-	// marshall to json
-	jsonPredicate, err := protojson.Marshal(predicate)
-	if err != nil {
-		return fmt.Errorf("marshalling predicate: %w", err)
-	}
-
-	// unmarshall to our typed predicate object
-	var p chainloop.ProvenancePredicateV02
-	err = json.Unmarshal(jsonPredicate, &p)
-	if err != nil {
-		return fmt.Errorf("unmarshalling predicate: %w", err)
-	}
-
-	// insert policy evaluations for attestation
-	if p.PolicyEvaluations == nil {
-		p.PolicyEvaluations = make(map[string][]*chainloop.PolicyEvaluation)
-	}
-	attEvaluations := make([]*chainloop.PolicyEvaluation, 0, len(policyResults))
-	for _, ev := range policyResults {
-		attEvaluations = append(attEvaluations, chainloop.RenderEvaluation(ev))
-	}
-	p.PolicyEvaluations[AttPolicyEvaluation] = attEvaluations
-
-	// marshall back to JSON
-	jsonPredicate, err = json.Marshal(p)
-	if err != nil {
-		return fmt.Errorf("marshalling predicate: %w", err)
-	}
-
-	// finally unmarshal from JSON to structpb.Struct.
-	var finalPredicate structpb.Struct
-	err = protojson.Unmarshal(jsonPredicate, &finalPredicate)
-	if err != nil {
-		return fmt.Errorf("unmarshalling predicate: %w", err)
-	}
-
-	statement.Predicate = &finalPredicate
-
-	return nil
 }
 
 func (ab *AttestationRenderer) envelopeToBundle(dsseEnvelope dsse.Envelope) (*protobundle.Bundle, error) {
